Removes the S3StorageProvider and StorageAccountManager dependencies on AuxRest and moves them into StringDataStoreFactory.
