+++ s3storageprovider/src/main/java/org/duracloud/s3storage/StringDataStoreFactory.java
@@ -7,12 +7,42 @@
  */
 package org.duracloud.s3storage;
 
+import org.duracloud.common.error.DuraCloudRuntimeException;
+import org.duracloud.storage.domain.StorageAccount;
+import org.duracloud.storage.domain.StorageAccountManager;
+
 /**
  * A simple factory for StringDataStore object.
+ *
  * @author Daniel Bernstein
  */
 public class StringDataStoreFactory {
-    public StringDataStore create(final String hiddenSpaceName, S3StorageProvider s3StorageProvider) {
-        return new StringDataStore(hiddenSpaceName, s3StorageProvider);
+
+    private StorageAccountManager storageAccountManager;
+
+    /**
+     * Default constructor
+     * @param storageAccountManager A storage account manager
+     */
+    public StringDataStoreFactory(StorageAccountManager storageAccountManager) {
+        this.storageAccountManager = storageAccountManager;
+    }
+
+    /**
+     * Create a StringDataStore
+     * @param hiddenSpaceName The name of the hidden space in which to store data
+     * @return
+     */
+    public StringDataStore create(final String hiddenSpaceName) {
+        return new StringDataStore(hiddenSpaceName, getS3StorageProvider());
+    }
+
+    private S3StorageProvider getS3StorageProvider() {
+        if (!storageAccountManager.isInitialized()) {
+            throw new DuraCloudRuntimeException("storageAccountManager is not initialized!!!");
+        }
+
+        final StorageAccount account = storageAccountManager.getPrimaryStorageAccount();
+        return new S3StorageProvider(account.getUsername(), account.getPassword(), account.getOptions());
     }
 }
